refactor: update lerobot import paths to remove 'common' prefix
