+++ physical_ai_server/physical_ai_server/inference/inference_manager.py
@@ -18,7 +18,7 @@
 
 import os
 
-from lerobot.common.policies.pretrained import PreTrainedPolicy
+from lerobot.policies.pretrained import PreTrainedPolicy
 import numpy as np
 from physical_ai_server.utils.read_file import read_json_file
 import torch
@@ -139,27 +139,27 @@ class InferenceManager:
 
     def _get_policy_class(self, name: str) -> PreTrainedPolicy:
         if name == 'tdmpc':
-            from lerobot.common.policies.tdmpc.modeling_tdmpc import TDMPCPolicy
+            from lerobot.policies.tdmpc.modeling_tdmpc import TDMPCPolicy
 
             return TDMPCPolicy
         elif name == 'diffusion':
-            from lerobot.common.policies.diffusion.modeling_diffusion import DiffusionPolicy
+            from lerobot.policies.diffusion.modeling_diffusion import DiffusionPolicy
 
             return DiffusionPolicy
         elif name == 'act':
-            from lerobot.common.policies.act.modeling_act import ACTPolicy
+            from lerobot.policies.act.modeling_act import ACTPolicy
 
             return ACTPolicy
         elif name == 'vqbet':
-            from lerobot.common.policies.vqbet.modeling_vqbet import VQBeTPolicy
+            from lerobot.policies.vqbet.modeling_vqbet import VQBeTPolicy
 
             return VQBeTPolicy
         elif name == 'pi0':
-            from lerobot.common.policies.pi0.modeling_pi0 import PI0Policy
+            from lerobot.policies.pi0.modeling_pi0 import PI0Policy
 
             return PI0Policy
         elif name == 'pi0fast':
-            from lerobot.common.policies.pi0fast.modeling_pi0fast import PI0FASTPolicy
+            from lerobot.policies.pi0fast.modeling_pi0fast import PI0FASTPolicy
             return PI0FASTPolicy
         # TODO: Uncomment when GrootN1Policy is implemented
         # elif name == 'groot-n1':
